introduce AbstractModelConverter as base for custom ModelConverters
