+++ modules/swagger-core/src/main/java/com/wordnik/swagger/jackson/ModelResolver.java
@@ -3,12 +3,9 @@ package com.wordnik.swagger.jackson;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElementWrapper;
@@ -16,19 +13,15 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.core.Version;
-import com.fasterxml.jackson.databind.AnnotationIntrospector;
 import com.fasterxml.jackson.databind.BeanDescription;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyMetadata;
-import com.fasterxml.jackson.databind.PropertyName;
-import com.fasterxml.jackson.databind.introspect.Annotated;
 import com.fasterxml.jackson.databind.introspect.AnnotatedMember;
 import com.fasterxml.jackson.databind.introspect.BeanPropertyDefinition;
 import com.fasterxml.jackson.databind.jsontype.NamedType;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.wordnik.swagger.annotations.ApiModel;
-import com.wordnik.swagger.annotations.ApiModelProperty;
 import com.wordnik.swagger.converter.ModelConverter;
 import com.wordnik.swagger.converter.ModelConverterContext;
 import com.wordnik.swagger.models.ComposedModel;
@@ -37,41 +30,16 @@ import com.wordnik.swagger.models.ModelImpl;
 import com.wordnik.swagger.models.RefModel;
 import com.wordnik.swagger.models.Xml;
 import com.wordnik.swagger.models.properties.ArrayProperty;
-import com.wordnik.swagger.models.properties.BooleanProperty;
-import com.wordnik.swagger.models.properties.DateTimeProperty;
-import com.wordnik.swagger.models.properties.DoubleProperty;
-import com.wordnik.swagger.models.properties.FloatProperty;
-import com.wordnik.swagger.models.properties.IntegerProperty;
-import com.wordnik.swagger.models.properties.LongProperty;
 import com.wordnik.swagger.models.properties.MapProperty;
 import com.wordnik.swagger.models.properties.Property;
 import com.wordnik.swagger.models.properties.RefProperty;
 import com.wordnik.swagger.models.properties.StringProperty;
 
-public class ModelResolver implements ModelConverter{
+public class ModelResolver extends AbstractModelConverter implements ModelConverter {
 	
-  protected final ObjectMapper _mapper;
-  protected final AnnotationIntrospector _intr;
-  protected final TypeNameResolver _typeNameResolver = TypeNameResolver.std;
-  
-
-  /**
-   * Minor optimization: no need to keep on resolving same types over and over
-   * again.
-   */
-  protected Map<JavaType, String> _resolvedTypeNames = new ConcurrentHashMap<JavaType, String>();
-  
   @SuppressWarnings("serial")
   public ModelResolver(ObjectMapper mapper) {
-    mapper.registerModule(
-      new SimpleModule("swagger", Version.unknownVersion()) {
-        @Override
-        public void setupModule(SetupContext context) {
-          context.insertAnnotationIntrospector(new SwaggerAnnotationIntrospector());
-        }
-      });
-    _mapper = mapper;
-    _intr = mapper.getSerializationConfig().getAnnotationIntrospector();
+	  super(mapper);
   }
 
   public ObjectMapper objectMapper() {
@@ -310,132 +278,5 @@ public class ModelResolver implements ModelConverter{
     model.setProperties(modelProps);
     context.defineModel(name, model);
     return model;
-  }
-
-  protected Property getPrimitiveProperty(String typeName) {
-    Property property = null;
-    if("boolean".equals(typeName)) {
-      property = new BooleanProperty();
-    }
-    if("string".equals(typeName)) {
-      property = new StringProperty();
-    }
-    else if("integer".equals(typeName)) {
-      property = new IntegerProperty();
-    }
-    else if("long".equals(typeName)) {
-      property = new LongProperty();
-    }
-    else if("float".equals(typeName)) {
-      property = new FloatProperty();
-    }
-    else if("double".equals(typeName)) {
-      property = new DoubleProperty();
-    }
-    else if("dateTime".equals(typeName)) {
-      property = new DateTimeProperty();
-    }
-    return property;
-  }
-
-/*  
-  protected void _addEnumProps(BeanPropertyDefinition propDef, Class<?> propClass,
-      ModelProperty result) {
-    final boolean useIndex =  _mapper.isEnabled(SerializationFeature.WRITE_ENUMS_USING_INDEX);
-    final boolean useToString = _mapper.isEnabled(SerializationFeature.WRITE_ENUMS_USING_TO_STRING);
-    List<AllowableValue> enums = new ArrayList<AllowableValue>();
-    @SuppressWarnings("unchecked")
-    Class<Enum<?>> enumClass = (Class<Enum<?>>) propClass;
-    for (Enum<?> en : enumClass.getEnumConstants()) {
-      String n;
-      if (useIndex) {
-        n = String.valueOf(en.ordinal());
-      } else if (useToString) {
-        n = en.toString();
-      } else {
-        n = _intr.findEnumValue(en);
-      }
-      enums.add(new AllowableValue(n));
-    }
-    result.setAllowableValues(enums);
-  }
-*/
-  protected String _description(Annotated ann) {
-    // while name suggests it's only for properties, should work for any Annotated thing.
-    // also; with Swagger introspector's help, should get it from ApiModel/ApiModelProperty
-    return _intr.findPropertyDescription(ann);
-  }
-
-  protected String _typeName(JavaType type) {
-    return _typeName(type, null);
-  }
-  
-  protected String _typeName(JavaType type, BeanDescription beanDesc) {
-    String name = _resolvedTypeNames.get(type);
-    if (name != null) {
-      return name;
-    }
-    name = _findTypeName(type, beanDesc);
-    _resolvedTypeNames.put(type,  name);
-    return name;
-  }
-
-  protected String _findTypeName(JavaType type, BeanDescription beanDesc) {
-    // First, handle container types; they require recursion
-    if (type.isArrayType())
-      return "Array";
-
-    if (type.isMapLikeType())
-      return "Map";
-
-    if (type.isContainerType()) {
-      if (Set.class.isAssignableFrom(type.getRawClass()))
-        return "Set";
-      return "List";
-    }
-    if (beanDesc == null) {
-      beanDesc = _mapper.getSerializationConfig().introspectClassAnnotations(type);
-    }
-      
-    PropertyName rootName = _intr.findRootName(beanDesc.getClassInfo());
-    if (rootName != null && rootName.hasSimpleName()) {
-      return rootName.getSimpleName();
-    }
-    return _typeNameResolver.nameForType(type);
-  }
-
-  protected String _typeQName(JavaType type) {
-    return type.getRawClass().getName();
-  }
-
-  protected String _subTypeName(NamedType type) {
-    // !!! TODO: should this use 'name' instead?
-    return type.getType().getName();
-  }
-
-  protected String _findExampleValue(Annotated a) {
-    ApiModelProperty prop = a.getAnnotation(ApiModelProperty.class);
-    if (prop != null) {
-      if (!prop.example().isEmpty()) {
-        return prop.example();
-      }
-    }
-    return null;
-  }
-
-  // TODO remove this
-  static Comparator<Property> getPropertyComparator() {
-    return new Comparator<Property>() {
-      @Override
-      public int compare(Property one, Property two) {
-        if (one.getPosition() == null && two.getPosition() == null)
-          return 0;
-        if (one.getPosition() == null)
-          return -1;
-        if (two.getPosition() == null)
-          return 1;
-        return one.getPosition().compareTo(two.getPosition());
-      }
-    };
   }
 }
