+++ modules/swagger-core/src/test/scala/com/wordnik/swagger/model/override/GenericModel.java
@@ -1,0 +1,37 @@
+package com.wordnik.swagger.model.override;
+
+import java.lang.reflect.Type;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
+public class GenericModel {
+	private static final Map<String,Class<?>> additionalPropertyTypes = new HashMap<String, Class<?>>();
+	
+	
+	public static void declareProperty(String key,Class<?> cls){
+		additionalPropertyTypes.put(key, cls);		
+	}
+	
+	public static Map<String,Class<?>> getDeclaredProperties(){
+		return Collections.unmodifiableMap(additionalPropertyTypes);
+	}
+	
+	public static void undeclareProperty(String key){
+		additionalPropertyTypes.remove(key);		
+	}
+	
+	
+	private Map<String,Object> additionalProperties;
+	
+	public void setProperty(String key, Object value){
+		if(value==null ||additionalPropertyTypes.get(key).isAssignableFrom(value.getClass())){
+			additionalProperties.put(key, value);
+		}
+	}
+	
+	public Object getProperty(String key){
+		return additionalProperties.get(key);
+	}
+	
+}
